tests/wpt/web-platform-tests/web-locks/secure-context.tentative.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web Locks API: API requires secure context</title>
<link rel=help href="https://wicg.github.io/web-locks/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(t => {
  assert_true(window.isSecureContext);
  assert_idl_attribute(navigator, 'locks',
                       'navigator.locks exists in secure context');
  assert_true('LockManager' in self,
              'LockManager is present in secure contexts');
  assert_true('Lock' in self,
              'Lock interface is present in secure contexts');
}, 'API presence in secure contexts');
</script>
